Update tests and python bindings (#10)
* Make fasttext test dependency optional

* Use long long instead of ssize_t in python bindings

* Add minimal serialization tests

* Add windows for python 3.10 to CI

+++ python/floret_module/floret/pybind/floret_pybind.cc
@@ -173,7 +173,7 @@ PYBIND11_MODULE(floret_pybind, m) {
       py::call_guard<py::gil_scoped_release>());
 
   py::class_<fasttext::Vector>(m, "Vector", py::buffer_protocol(), py::module_local())
-      .def(py::init<ssize_t>())
+      .def(py::init<long long>())
       .def_buffer([](fasttext::Vector& m) -> py::buffer_info {
         return py::buffer_info(
             m.data(),
@@ -187,7 +187,7 @@ PYBIND11_MODULE(floret_pybind, m) {
   py::class_<fasttext::DenseMatrix>(
       m, "DenseMatrix", py::buffer_protocol(), py::module_local())
       .def(py::init<>())
-      .def(py::init<ssize_t, ssize_t>())
+      .def(py::init<long long, long long>())
       .def_buffer([](fasttext::DenseMatrix& m) -> py::buffer_info {
         return py::buffer_info(
             m.data(),
